feat: don't allow admins to leave the community
