+++ packages/workers/communities/src/handlers/post/staffPickCommunity.ts
@@ -5,7 +5,7 @@ import { error, type IRequest } from 'itty-router';
 import { Client } from 'pg';
 import { object, string } from 'zod';
 
-import type { Env } from '../types';
+import type { Env } from '../../types';
 
 type ExtensionRequest = {
   id: string;
